refine join and leave broadcast
